EE採掘機実装
テクスチャ未作成
EEオファレン砲台はレーザーの方向を直せば完成

+++ src/main/java/nahamawiki/oef/tileentity/ITileEntityEEMachine.java
@@ -23,6 +23,6 @@ public interface ITileEntityEEMachine {
 	/** 機械の情報を返す。 */
 	public String[] getState();
 
-	public int getLevel(int meta);
+	public byte getLevel(int meta);
 
 }
